Modify bash script
